# internal/docs/swagger.yaml
definitions:
...
  dto.EditorCreateRequest:
    properties:
      is_dir:
        type: boolean
      path:
        type: string
    required:
    - path
    type: object
  dto.EditorDeleteRequest:
    properties:
      path:
        type: string
    required:
    - path
    type: object
  dto.EditorFileInfo:
    properties:
      is_dir:
        type: boolean
      modified:
        type: string
      name:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
  dto.EditorFileRequest:
    properties:
      file_path:
        type: string
    required:
    - file_path
    type: object
  dto.EditorFileResponse:
    properties:
      content:
        type: string
      file_path:
        type: string
      language:
        description: File language/extension
        type: string
      size:
        type: integer
    type: object
  dto.EditorListRequest:
    properties:
      directory_path:
        type: string
      max_depth:
        type: integer
      recursive:
        type: boolean
    required:
    - directory_path
    type: object
  dto.EditorListResponse:
    properties:
      directory_path:
        type: string
      files:
        items:
          $ref: '#/definitions/dto.EditorFileInfo'
        type: array
      total_files:
        type: integer
    type: object
  dto.EditorRenameRequest:
    properties:
      new_path:
        type: string
      old_path:
        type: string
    required:
    - new_path
    - old_path
    type: object
  dto.EditorSaveRequest:
    properties:
      content:
        type: string
      file_path:
        type: string
    required:
    - content
    - file_path
    type: object
  dto.EditorSaveResponse:
    properties:
      file_path:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /api/v1/editor/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Path and type
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditorCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorSaveResponse'
      summary: Create file/directory
      tags:
      - Editor
  /api/v1/editor/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Path to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditorDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorSaveResponse'
      summary: Delete file/directory
      tags:
      - Editor
  /api/v1/editor/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Directory path
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditorListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorListResponse'
      summary: List files
      tags:
      - Editor
  /api/v1/editor/read:
    post:
      consumes:
      - application/json
      parameters:
      - description: File path
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditorFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorFileResponse'
      summary: Read file
      tags:
      - Editor
  /api/v1/editor/rename:
    post:
      consumes:
      - application/json
      parameters:
      - description: Old and new paths
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditorRenameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorSaveResponse'
      summary: Rename/move file/directory
      tags:
      - Editor
  /api/v1/editor/save:
    post:
      consumes:
      - application/json
      parameters:
      - description: File path and content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditorSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorSaveResponse'
      summary: Save file
      tags:
      - Editor
...
  /backup/export:
    get:
      description: Exports all tables to JSON and creates a downloadable zip file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Export full database backup
      tags:
      - Backup
  /backup/import:
    post:
      consumes:
      - multipart/form-data
      description: Restore data from a backup zip file
      parameters:
      - description: Backup zip file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Import database backup
      tags:
      - Backup
...
  /llm/status:
    get:
      description: Check if Ollama is installed and list available models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get local LLM status
      tags:
      - LLM
  /llm/test:
    post:
      consumes:
      - application/json
      description: Send a test message to Ollama (if available) or fallback to Claude
        API
      parameters:
      - description: Test request
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Test local LLM inference
      tags:
      - LLM
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token (e.g., "Bearer
      <token>")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"